requirements.txt
python-telegram-bot[job-queue]>=20.0
pymongo>=4.0.0
python-dotenv>=0.19.0
pyrogram>=2.0.0,<3.0.0
